+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 
+## 0.98.6
+
 ## 0.98.5
 - Multiline Curl commands are now supported
 - `to_curl/2` now supports short and long argument flag generation